--- implementacao/exemplo_linear_elasticity.py
@@ -1,58 +1,0 @@
-# Scaled variable
-L = 1
-W = 0.2
-mu = 1
-rho = 1
-delta = W/L
-gamma = 0.4*delta**2
-beta = 1.25
-lambda_ = beta
-g = gamma
-
-import numpy as np
-import ufl
-
-from mpi4py import MPI
-from petsc4py.PETSc import ScalarType
-
-from dolfinx import mesh, fem, plot, io
-
-domain = mesh.create_box(MPI.COMM_WORLD, [np.array([0,0,0]), np.array([L, W, W])],
-                  [20,6,6], cell_type=mesh.CellType.hexahedron)
-V = fem.VectorFunctionSpace(domain, ("CG", 1))
-
-
-
-#Condições de contorno 
-def clamped_boundary(x):
-    return np.isclose(x[0], 0)
-
-fdim = domain.topology.dim - 1
-boundary_facets = mesh.locate_entities_boundary(domain, fdim, clamped_boundary)
-u_D = np.array([0,0,0], dtype=ScalarType)
-bc = fem.dirichletbc(u_D, fem.locate_dofs_topological(V, fdim, boundary_facets), V)
-T = fem.Constant(domain, ScalarType((0, 0, 0)))
-ds = ufl.Measure("ds", domain=domain)
-
-"""Formulação variacional"""
-def epsilon(u): #deformação
-    return ufl.sym(ufl.grad(u)) # Equivalent to 0.5*(ufl.nabla_grad(u) + ufl.nabla_grad(u).T)
-def sigma(u): #tensao
-    return lambda_ * ufl.nabla_div(u) * ufl.Identity(len(u)) + 2*mu*epsilon(u)
-
-
-
-
-u = ufl.TrialFunction(V)
-v = ufl.TestFunction(V)
-f = fem.Constant(domain, ScalarType((0, 0, -rho*g)))
-a = ufl.inner(sigma(u), epsilon(v)) * ufl.dx
-L = ufl.dot(f, v) * ufl.dx + ufl.dot(T, v) * ds
-
-
-problem = fem.petsc.LinearProblem(a, L, bcs=[bc], petsc_options={"ksp_type": "preonly", "pc_type": "lu"})
-uh = problem.solve()
-
-"""Calculo do tensoes von mises"""
-s = sigma(uh) -1./3* ufl.tr(sigma(uh)) *ufl.Identity(len(uh))
-von_Mises = ufl.sqrt(3./2*ufl.inner(s, s))
