--- resultados/exemplo_7_3.xdmf
@@ -1,32 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE Xdmf SYSTEM "Xdmf.dtd" []>
-<Xdmf Version="3.0" xmlns:xi="http://www.w3.org/2001/XInclude">
-  <Domain>
-    <Grid Name="mesh" GridType="Uniform">
-      <Topology TopologyType="Quadrilateral" NumberOfElements="4332" NodesPerElement="4">
-        <DataItem Dimensions="4332 4" NumberType="Int" Format="HDF">exemplo_7_3.h5:/Mesh/mesh/topology</DataItem>
-      </Topology>
-      <Geometry GeometryType="XYZ">
-        <DataItem Dimensions="4466 3" Format="HDF">exemplo_7_3.h5:/Mesh/mesh/geometry</DataItem>
-      </Geometry>
-    </Grid>
-    <Grid Name="Facet tags" GridType="Uniform">
-      <xi:include xpointer="xpointer(/Xdmf/Domain/Grid/Geometry)" />
-      <Topology TopologyType="PolyLine" NumberOfElements="266" NodesPerElement="2">
-        <DataItem Dimensions="266 2" NumberType="Int" Format="HDF">exemplo_7_3.h5:/MeshTags/Facet tags/topology</DataItem>
-      </Topology>
-      <Attribute Name="Facet tags" AttributeType="Scalar" Center="Cell">
-        <DataItem Dimensions="266 1" Format="HDF">exemplo_7_3.h5:/MeshTags/Facet tags/Values</DataItem>
-      </Attribute>
-    </Grid>
-    <Grid Name="Cell tags" GridType="Uniform">
-      <xi:include xpointer="xpointer(/Xdmf/Domain/Grid/Geometry)" />
-      <Topology TopologyType="Quadrilateral" NumberOfElements="4332" NodesPerElement="4">
-        <DataItem Dimensions="4332 4" NumberType="Int" Format="HDF">exemplo_7_3.h5:/MeshTags/Cell tags/topology</DataItem>
-      </Topology>
-      <Attribute Name="Cell tags" AttributeType="Scalar" Center="Cell">
-        <DataItem Dimensions="4332 1" Format="HDF">exemplo_7_3.h5:/MeshTags/Cell tags/Values</DataItem>
-      </Attribute>
-    </Grid>
-  </Domain>
-</Xdmf>
