--- resultados/exemplo_linear_elasticity.xdmf
@@ -1,14 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE Xdmf SYSTEM "Xdmf.dtd" []>
-<Xdmf Version="3.0" xmlns:xi="http://www.w3.org/2001/XInclude">
-  <Domain>
-    <Grid Name="mesh" GridType="Uniform">
-      <Topology TopologyType="Triangle" NumberOfElements="160000" NodesPerElement="3">
-        <DataItem Dimensions="160000 3" NumberType="Int" Format="HDF">exemplo_linear_elasticity.h5:/Mesh/mesh/topology</DataItem>
-      </Topology>
-      <Geometry GeometryType="XY">
-        <DataItem Dimensions="80601 2" Format="HDF">exemplo_linear_elasticity.h5:/Mesh/mesh/geometry</DataItem>
-      </Geometry>
-    </Grid>
-  </Domain>
-</Xdmf>
