--- resultados/malha_refi.xdmf
@@ -1,41 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE Xdmf SYSTEM "Xdmf.dtd" []>
-<Xdmf Version="3.0" xmlns:xi="http://www.w3.org/2001/XInclude">
-  <Domain>
-    <Grid Name="mesh" GridType="Uniform">
-      <Topology TopologyType="Quadrilateral" NumberOfElements="5304" NodesPerElement="4">
-        <DataItem Dimensions="5304 4" NumberType="Int" Format="HDF">malha_refi.h5:/Mesh/mesh/topology</DataItem>
-      </Topology>
-      <Geometry GeometryType="XY">
-        <DataItem Dimensions="5480 2" Format="HDF">malha_refi.h5:/Mesh/mesh/geometry</DataItem>
-      </Geometry>
-    </Grid>
-    <Grid Name="Facet tags" GridType="Uniform">
-      <xi:include xpointer="xpointer(/Xdmf/Domain/Grid/Geometry)" />
-      <Topology TopologyType="PolyLine" NumberOfElements="175" NodesPerElement="2">
-        <DataItem Dimensions="175 2" NumberType="Int" Format="HDF">malha_refi.h5:/MeshTags/Facet tags/topology</DataItem>
-      </Topology>
-      <Attribute Name="Facet tags" AttributeType="Scalar" Center="Cell">
-        <DataItem Dimensions="175 1" Format="HDF">malha_refi.h5:/MeshTags/Facet tags/Values</DataItem>
-      </Attribute>
-    </Grid>
-    <Grid Name="Cell tags" GridType="Uniform">
-      <xi:include xpointer="xpointer(/Xdmf/Domain/Grid/Geometry)" />
-      <Topology TopologyType="Quadrilateral" NumberOfElements="5304" NodesPerElement="4">
-        <DataItem Dimensions="5304 4" NumberType="Int" Format="HDF">malha_refi.h5:/MeshTags/Cell tags/topology</DataItem>
-      </Topology>
-      <Attribute Name="Cell tags" AttributeType="Scalar" Center="Cell">
-        <DataItem Dimensions="5304 1" Format="HDF">malha_refi.h5:/MeshTags/Cell tags/Values</DataItem>
-      </Attribute>
-    </Grid>
-    <Grid Name="f" GridType="Collection" CollectionType="Temporal">
-      <Grid Name="f" GridType="Uniform">
-        <xi:include xpointer="xpointer(/Xdmf/Domain/Grid[@GridType='Uniform'][1]/*[self::Topology or self::Geometry])" />
-        <Time Value="0" />
-        <Attribute Name="f" AttributeType="Scalar" Center="Node">
-          <DataItem Dimensions="5480 1" Format="HDF">malha_refi.h5:/Function/f/0</DataItem>
-        </Attribute>
-      </Grid>
-    </Grid>
-  </Domain>
-</Xdmf>
