--- resultados/malha_refin.xdmf
@@ -1,14 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE Xdmf SYSTEM "Xdmf.dtd" []>
-<Xdmf Version="3.0" xmlns:xi="http://www.w3.org/2001/XInclude">
-  <Domain>
-    <Grid Name="mesh" GridType="Uniform">
-      <Topology TopologyType="PolyLine" NumberOfElements="175" NodesPerElement="2">
-        <DataItem Dimensions="175 2" NumberType="Int" Format="HDF">malha_refin.h5:/Mesh/mesh/topology</DataItem>
-      </Topology>
-      <Geometry GeometryType="XY">
-        <DataItem Dimensions="176 2" Format="HDF">malha_refin.h5:/Mesh/mesh/geometry</DataItem>
-      </Geometry>
-    </Grid>
-  </Domain>
-</Xdmf>
