--- resultados/malha_refinada.xdmf
@@ -1,14 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE Xdmf SYSTEM "Xdmf.dtd" []>
-<Xdmf Version="3.0" xmlns:xi="http://www.w3.org/2001/XInclude">
-  <Domain>
-    <Grid Name="mesh" GridType="Uniform">
-      <Topology TopologyType="Triangle" NumberOfElements="1600" NodesPerElement="3">
-        <DataItem Dimensions="1600 3" NumberType="Int" Format="HDF">malha_refinada.h5:/Mesh/mesh/topology</DataItem>
-      </Topology>
-      <Geometry GeometryType="XY">
-        <DataItem Dimensions="861 2" Format="HDF">malha_refinada.h5:/Mesh/mesh/geometry</DataItem>
-      </Geometry>
-    </Grid>
-  </Domain>
-</Xdmf>
