--- resultados/mesh_out.xdmf
@@ -1,32 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE Xdmf SYSTEM "Xdmf.dtd" []>
-<Xdmf Version="3.0" xmlns:xi="http://www.w3.org/2001/XInclude">
-  <Domain>
-    <Grid Name="mesh" GridType="Uniform">
-      <Topology TopologyType="Tetrahedron" NumberOfElements="192674" NodesPerElement="4">
-        <DataItem Dimensions="192674 4" NumberType="Int" Format="HDF">mesh_out.h5:/Mesh/mesh/topology</DataItem>
-      </Topology>
-      <Geometry GeometryType="XYZ">
-        <DataItem Dimensions="36163 3" Format="HDF">mesh_out.h5:/Mesh/mesh/geometry</DataItem>
-      </Geometry>
-    </Grid>
-    <Grid Name="Facet tags" GridType="Uniform">
-      <xi:include xpointer="xpointer(/Xdmf/Domain/Grid/Geometry)" />
-      <Topology TopologyType="Triangle" NumberOfElements="20442" NodesPerElement="3">
-        <DataItem Dimensions="20442 3" NumberType="Int" Format="HDF">mesh_out.h5:/MeshTags/Facet tags/topology</DataItem>
-      </Topology>
-      <Attribute Name="Facet tags" AttributeType="Scalar" Center="Cell">
-        <DataItem Dimensions="20442 1" Format="HDF">mesh_out.h5:/MeshTags/Facet tags/Values</DataItem>
-      </Attribute>
-    </Grid>
-    <Grid Name="Cell tags" GridType="Uniform">
-      <xi:include xpointer="xpointer(/Xdmf/Domain/Grid/Geometry)" />
-      <Topology TopologyType="Tetrahedron" NumberOfElements="192674" NodesPerElement="4">
-        <DataItem Dimensions="192674 4" NumberType="Int" Format="HDF">mesh_out.h5:/MeshTags/Cell tags/topology</DataItem>
-      </Topology>
-      <Attribute Name="Cell tags" AttributeType="Scalar" Center="Cell">
-        <DataItem Dimensions="192674 1" Format="HDF">mesh_out.h5:/MeshTags/Cell tags/Values</DataItem>
-      </Attribute>
-    </Grid>
-  </Domain>
-</Xdmf>
