--- resultados/poisson_ex1_neumann_and_dirichlet.xdmf
@@ -1,23 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE Xdmf SYSTEM "Xdmf.dtd" []>
-<Xdmf Version="3.0" xmlns:xi="http://www.w3.org/2001/XInclude">
-  <Domain>
-    <Grid Name="mesh" GridType="Uniform">
-      <Topology TopologyType="Quadrilateral" NumberOfElements="475212" NodesPerElement="4">
-        <DataItem Dimensions="475212 4" NumberType="Int" Format="HDF">poisson_ex1_neumann_and_dirichlet.h5:/Mesh/mesh/topology</DataItem>
-      </Topology>
-      <Geometry GeometryType="XY">
-        <DataItem Dimensions="476606 2" Format="HDF">poisson_ex1_neumann_and_dirichlet.h5:/Mesh/mesh/geometry</DataItem>
-      </Geometry>
-    </Grid>
-    <Grid Name="f" GridType="Collection" CollectionType="Temporal">
-      <Grid Name="f" GridType="Uniform">
-        <xi:include xpointer="xpointer(/Xdmf/Domain/Grid[@GridType='Uniform'][1]/*[self::Topology or self::Geometry])" />
-        <Time Value="0" />
-        <Attribute Name="f" AttributeType="Scalar" Center="Node">
-          <DataItem Dimensions="476606 1" Format="HDF">poisson_ex1_neumann_and_dirichlet.h5:/Function/f/0</DataItem>
-        </Attribute>
-      </Grid>
-    </Grid>
-  </Domain>
-</Xdmf>
