--- scripts/__init__.py
@@ -1,1 +1,0 @@
-from .operacoes_02 import Hencky_Strain
