--- tests/test_ex001.py
@@ -1,7 +1,0 @@
-from ..Exemplos.ex001 import carregamento
-from unittest import*
-
-def test_meu_primeiro_test():
-    assert True
-    carregamento == 50000
-
